// utils/types.ts
export type UserRole = 'administrator' | 'operator_unit' | 'validator' | 'verifikator';

export type ActiveMenu =
  | 'dashboard'
  | 'master'
  | 'master_renstra'
  | 'pengaturan_kinerja'
  | 'input_kinerja'
  | 'capaian_bulan'
  | 'capaian_triwulan'
  | 'lhe';

export interface User {
  id: string;
  name: string;
  nip: string;
  email: string;
  role: UserRole;
  roleTitle: string;
  opdId: string;
  opdName: string;
  avatarUrl?: string;
}

export interface OPD {
  id: string;
  kode: string;
  nama: string;
  kepala: string;
  nipKepala: string;
  kategori: string;
  email: string;
  telepon: string;
}

export interface ProgramKegiatan {
  id: string;
  opdId: string;
  kodeProgram: string;
  namaProgram: string;
  kodeKegiatan: string;
  namaKegiatan: string;
  paguAnggaran: number;
  tahun: number;
}

export interface RenstraTujuan {
  id: string;
  kode: string;
  pernyataan: string;
  indikator: string;
  targetAwal: string;
  targetAkhir: string;
  satuan: string;
}

export interface RenstraSasaran {
  id: string;
  opdId: string;
  tujuanId: string;
  kode: string;
  sasaranStrategis: string;
  indikatorKinerja: string;
  satuan: string;
  targetTahun1: number;
  targetTahun2: number;
  targetTahun3: number;
  targetTahun4: number;
  targetTahun5: number;
  cascadingLevel: 'Eselon II' | 'Eselon III' | 'Eselon IV' | 'Staf/Pelaksana';
}

export type Polarisasi = 'Maximize' | 'Minimize';

export interface IndikatorPK {
  id: string;
  opdId: string;
  tahun: number;
  sasaranStrategis: string;
  namaIndikator: string;
  formula: string;
  satuan: string;
  polarisasi: Polarisasi;
  targetTahunan: number;
  targetT1: number;
  targetT2: number;
  targetT3: number;
  targetT4: number;
  paguAnggaran: number;
  penanggungJawab: string;
  tipeIndikator: 'IKU' | 'IKP' | 'Program' | 'Kegiatan';
}